BYT: Run surfaceflinger as a separate process
BZ: 91727

Depends on:
    vendor/intel/common - Change-Id I03edbc9b54b150afc24b2426b5e675887bb50f6f

Symptom: For Baytrail, boot animation sometimes hangs.

Diagnosis: Deadly embrace: The surfaceflinger thread within system_server
is waiting for a lock that is held by an OpenGL application (from within
the OpenGL library).  At the same time, that OpenGL application is blocked
waiting for surfaceflinger to give it a buffer.

Credit for the above diagnosis:
    Gary K Smith <gary.k.smith@intel.com>
    Piotr Luc <piotr.luc@intel.com>

Solution:

This is a workaround.  A true fix should resolve the locking issue.

This problem only occurs when surfaceflinger is run as a thread
in system_server, and not when surfaceflinger is run as a separate
process.

For Baytrail, run surfaceflinger as a separate process.

In init.baylake.rc:
- setprop system_init.startsurfaceflinger 0
- Define service surfaceflinger

Change-Id: Ia1b58b377e2fab07fdbf498d777c362f2c4cb211
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/96478
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>