OSIP installer: Fix invalid block device check
BZ: 118937

We were checking for the presence of a file to mark the error.

And as none of the block devices will be a valid file, the condition
will never enter.

Let's fix it by doing the proper check for a missing presence for a
block device, which is the intended check.

Change-Id: Ie79fe53edc4965100b44964cb86ce68539cefb93
Signed-off-by: Sergio Aguirre <sergio.a.aguirre.rodriguez@intel.com>
Signed-off-by: Frederic Despalles <frederic.fespalles@intel.com>
Reviewed-on: http://android.intel.com:8080/127981
Reviewed-by: Fradin, Benoit <benoit.fradin@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>