BYT: Again execute surfaceflinger inside system_server
BZ: 120731

This reverts commit 6106c96ac76dc82af24bf151b68fd43939a9f181,
which was entitled "BYT: Run surfaceflinger as a separate process".
It also removes similar changes made for other BYT targets.

The original commit (now being reverted) caused surfaceflinger to
run as a separate process instead of running as a thread inside
system_server.  The reverted commit was a workaround intended to be
used until a real solution to a locking issue was implemented.

Change-Id: I77cf58f571aaa9704d43a3f4d95bc88f7e98b108
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/116651
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>