BYT: Again (redux) execute surfaceflinger inside system_server
BZ: 120731

This is the second attempt to move surfaceflinger back inside
system_server.  The first attempt was reverted due to problems that
needed to be fixed elsewhere.

The original commit (now being reversed) caused surfaceflinger to
run as a separate process instead of running as a thread inside
system_server and was a workaround intended to be used until a real
solution to a locking issue was implemented.

Previous related changes:
http://android.intel.com:8080/116651
http://android.intel.com:8080/118437

Change-Id: I7524e84708188c92c314602a7a9e2d2e1f91b1ff
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/122845
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>