[Audio] Remove HwCodeComm route from FFRD8 routing files
BZ: 121353

The route HwCodecComm does not exist on FFRD8 as VoIP is going
through media path. Because of the presence of HwCodecComm route
in the PFW routing domains, the PFW is not able to start.

This patch removes rules that contains reference to the route
HwCodecComm from the routing files.

Change-Id: I6b5b9858e27c7324b32d334aa5202d707e9b717d
Signed-off-by: Guillaume Denneulin <guillaume.denneulin@intel.com>
Reviewed-on: http://android.intel.com:8080/117041
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Boisnard, FredericX <fredericx.boisnard@intel.com>
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>