[AUDIO] Set audio modem default volume gains to 0dB and swap I2S1/I2S2 modem configurations
BZ: 122988

Today, all the volumes are muted inside the modem.
We must set them to 0dB by default.

Moreover, on FFRD8, modem port I2S1 is used for mixing and modem port I2S2 is used for voice.
It was the other way around on CTP-RhB.
Today, the audio HAL is applying the CTP-RhB configuration.
So we need to swap the current configuration.

Change-Id: I949876a99056ea18aced407cc31302d80d961997
Signed-off-by: Neil Di Folco <neil.di.folco@intel.com>
Reviewed-on: http://android.intel.com:8080/118261
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>