charger: wait for fg_conf to be available before invoking
BZ: 121922

When mount_all function returns it is not guaranteed that partitions
are actually mounted, and as on charger mode event post-fs-data is
not available to insure partition are mounted, invoking fg_conf after
mount_all is not safe. So wait command on the executable file path
is used before starting service to make sure we can safely call it.

Change-Id: I409e0ed9d7f0377623bffb315c9a49a4b43864ca
Signed-off-by: Gabriel Touzeau <gabrielx.touzeau@intel.com>
Reviewed-on: http://android.intel.com:8080/119197
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Auger, Florent <florent.auger@intel.com>
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>