EM/THERMAL: Add ErrorCorrection tag in xml
BZ: 126732

This patch adds an 'ErrorCorrection' tag in
thermal_sensor_config.xml. This is used only
when a sensor supports interrupt mode.

The ErrorCorrection temperature is defined
as the maximum difference between the threshold
configured in the XML and the threshold
programmed in the sysfs. After characterization,
this value is found to be no more than 400mC,
for BYT T FFRD10 platform, with existing
slope/intercept values.

Hence, for safety, we are configuring
2 * max difference = 800mC, in this XML.

Change-Id: Ie5d511aef3392f94885bbfac656a51fac8bb1c21
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Reviewed-on: http://android.intel.com:8080/122327
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Roychowdhury, Saikat <saikat.roychowdhury@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>