[PORT FROM MAIN][BYT] Set IAFW component name in BoardConfig.mk
BZ: 126065

The IAFW component name may be different per board, e.g.
currently the code for BYT-T and BYT-M is different and
located within different subdirectories.

Add a new vaviable BOARD_IAFW_COMPONENT to specify the
component to use for the given board and set that appropriately
for each Baytrail board currently known.

This variable should be set with the name of the component to
build in "byt-iafw" for that board.

Change-Id: I39165ad2321141a7180a6009368aa9d21c23d51b
Orig-Change-Id: I99a4f1f58c7f00b02094beed0ec545cc8cbef49c
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>
Reviewed-on: http://android.intel.com:8080/133019
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>