Explicitly select FFRD8 panel, as default is no longer appropriate
BZ: 135734

Explicitly select the proper panel for FFRD8 and adjust comments to
match the current reality with respect to panel defaults.

A recent change has made the default Baytrail panel be eDP.  Before,
the default was whichever was present of eDP or PANASONIC_VXX09F006A00.
(Someday, the VBT will provide this information and build-time
selection will not be needed).

The change in default panel selection caused FFRD8 to stop working,
as its panel was not, as was newly required, explicitly selected.

The changes that modified the selection of the default panel:
R42B - http://android.intel.com:8080/127703
MAIN - http://android.intel.com:8080/128408

Change-Id: Ie4dd128132e3e27a95aaf174f1e65fb553f61308
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/129675
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>