watchdog: recovery, droidboot and charger mode should reset the wdt counter
BZ: 136727

The property sys.boot_completed is only set in MOS by the Android
framework. The other OS and the charger mode can't use this property
change as the event that tells them they can clean the osnib watchdog
counter. They have to detect that their main process has been
successfully started and is running to decide to clear the watchdog
counter. This patch exploits the init.svc.<process> properties to do
so.

Change-Id: Id91e8c7b0e0110c22f1c8e44fbf32291fa8e28da
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: http://android.intel.com:8080/131413
Reviewed-by: Fradin, Benoit <benoit.fradin@intel.com>
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Trebosc, FrancoisX <francoisx.trebosc@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>