[PORT FROM R42B-STABLE] watchdog: recovery, droidboot and charger mode should reset the wdt counter
BZ: 136727

The property sys.boot_completed is only set in MOS by the Android
framework. The other OS and the charger mode can't use this property
change as the event that tells them they can clean the osnib watchdog
counter. They have to detect that their main process has been
successfully started and is running to decide to clear the watchdog
counter. This patch exploits the init.svc.<process> properties to do
so.

Change-Id: I4efd875848ec54cb9ad689ec1ad2c17298fb2803
Orig-Change-Id: Id91e8c7b0e0110c22f1c8e44fbf32291fa8e28da
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Signed-off-by: Benoit Fradin <benoit.fradin@intel.com>
Reviewed-on: http://android.intel.com:8080/133091
Reviewed-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>