[K3.10] REVERTME - gfx - temp k3.10 baylake - setprop intel.hwc.overlayrgb.enable 0
BZ: 141817

This workaround was previously done for FFRD8 and needs to be done
for baylake RVP also.  This (along with other things) will allow mouse
cursor, plus home and back icons to display.

This commit only touches files that are specific to baylake_next:
    modified:   baylake/baylake_next.mk
    new file:   baylake/init.baylake_next.gengfx.rc

As a temporary measure for baylake during port to kernel 3.10 of Baytrail
graphics patches, set the following for baylake_next only:

setprop intel.hwc.overlayrgb.enable 0

Change-Id: I4aed1ef8efedcc72515e2942bcdf0179f20c530b
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/135020
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>