[AUDIO] Configure REALTEK Internal Muxes to bypass Audio DSP using PFW.
BZ: 140728

Realtek default MIC audio path is using a DSP.
This DSP block is now switched off to save power consumption.
Audio coming from MIC is thus very poor and cut.
This is resulting in bad uplink audio in CSV call.

This patch is configuring Realtek internal muxes so that
the audio stream is bypassing the DSP.

Change-Id: Ibdafc16ddabf80fae188a0c271b2e0bd6db95f19
Signed-off-by: Cedric Ducottet <cedricX.ducottet@intel.com>
Reviewed-on: http://android.intel.com:8080/135533
Reviewed-by: Ducottet, CedricX <cedricx.ducottet@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>