byt-t-ffrd10: set the default snapshot size to 1080p
BZ: 140905

Root cause:
Barcodes scanner doesn't care about the size of snapshot,
it use the default value that hal set.
And it will set the preview size which is matched with
screen size.
Our sensor can only output one ratio, so to choose
a ratio to match the screen is the best way to avoid stretch.

Change-Id: I4a5e77021af34cab4690e24ee0a6c372a12b973f
Signed-off-by: Yu CUI <yux.cui@intel.com>