byt-t-ffrd8: set the default snapshot size to 1080p
BZ: 149666

Root cause:
Barcodes scanner doesn't care about the size of snapshot,
it uses the default value which is set in camera hal.
And it will set the preview size which is matched with
screen size.
Our sensor can only output one ratio, so to choose
a ratio to match the screen is the best way to avoid stretch.

Change-Id: If12d1a7bc2d41e4d6b2cfd33967f6c26ff437076
Signed-off-by: Yu CUI <yux.cui@intel.com>