[AUDIO][Baylake] PFW XML Generation at Build Time
BZ: 150632

For the moment, .pfw routing files are not automatically generated
at build time on Baylake.

This patch enables the generation at build time of the XML file
containing the audio routing. This file is generated in the out/
directory under the name "AudioConfigurableDomains.xml".
To generate the file by hand, use the following command:
  m parameter-framework.audio.baylake

Change-Id: Ic4bd76b157e3f324132163b80dbd9a39873be80d
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>