[AUDIO][FFRD8] PFW XML Generation at Build Time
BZ: 152224

For the moment, .pfw routing files are not automatically generated
at build time on Baytrail/ffrd8.

This patch enables the generation at build time of the XML file
containing the audio routing. This file is generated in the out/
directory under the name "AudioConfigurableDomains.xml".
To generate the file by hand, use the following command:
  m parameter-framework.audio.byt_t_ffrd8

Change-Id: I5870f84cff198f0726d4641e84037da4e9c7c619
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>