Update ro.opengles.version to report GLES 3.0
BZ:152464

GLES 3.0 is supported by the UFO driver in JB MR2.  So the property
ro.opengles.version should be updated to 196608 (= 0x30000) to report
the correct GLES version. This fixes a failure in the GCTS test
android.graphics#testOpenGlEsVersion and also allows the GLES 3.0
version of the 3DRating benchmark to run.

Change-Id: I8a75a34a7c92d1613241d8ee66a7180dc2ce1474
Signed-off-by: Chris Harris <chris.harris@intel.com>