Change mode of sensorhubd socket to 666
BZ: 152356

The bug is, after switching to normal user other than owner,
sensor can't be calibrated anymore.

The root cause is, the normal user does not belong to system
group, thus does not have the privilege to access sensorhubd
socket.

The fix is, change sensorhubd socket from 600 to 666. So that
normal user can access it.

Change-Id: Icbff2158163e5d44761bfda3ca68ba4dd7ea11e1
Signed-off-by: Forrest Zhao <forrest.zhao@intel.com>