[AUDIO][BYT_T_CRV2] PFW XML Generation at Build Time
BZ: 158340

For the moment, .pfw routing files are not automatically generated
at build time on Baytrail/byt_t_crv2.

This patch enables the generation at build time of the XML file
containing the audio routing. This file is generated in the out/
directory under the name "AudioConfigurableDomains.xml".
To generate the file by hand, use the following command:
  m parameter-framework.audio.byt_t_crv2

Change-Id: Ic10e8e681befa3e76774a6ce44390a59245c2c1f
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>