byt_t_ffrd8: remove lte_wcdma_phone and use config_enabled_lte
BZ: 158931

lte_wcdma_phone was introduced when google was not
showing the 4G or LTE option for devices which support
LTE and WCDMA.

Since KK, 4G or LTE option is shown for devices which
support LTE and WCDMA, remove the config introduced
and use the google provided config flags.

Also, files and folders which are introduced for overriding
preferred_network_mode_choices and preferred_network_mode_values
are removed.

Change-Id: Id033b2d35f4092f017b98d883f0d70e1da697e69
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>