[AUDIO] Change audio stream mute/unmute sequence
BZ: 158292

In some cases, alsa mixer controls are OFF while the audio stream
is active.
This makes audio to be barely audible.
This is because the PFW uses the wrong alsa mixer controls to
mute/unmute the audio stream during the audio routing sequence.

This patch provides the right alsa mixer to mute/unmute audio stream.
This patch also provides the alsa mixer controls to handle
independently input volume and output volume.

Change-Id: Ib50f779161a73484faea43071c3db72ccc3d04f5
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>