init.byt.gengfx.rc--Create /mnt/libreq mountpoint for coreu service
BZ: 161950

Serivce coreu needs to mount a tmpfs as foundation for an efficient shared memory
implementation. It is difficult to create the mount point at runtime as "/" is
mounted as read-only. Therefore, create mount point during init process before
"/" is remounted as read only. This is done in the same file as the definition
of coreu service.

Change-Id: I76711ea6143954b9336cc32fd0975041561f3db8
Signed-off-by: Yunwei Zhang <yunwei.zhang@intel.com>