init.byt.gengfx.rc--Create /mnt/libreq mountpoint for coreu service
BZ: 161954

Serivce coreu needs to mount a tmpfs as foundation for an efficient shared memory
implementation. It is difficult to create the mount point at runtime as "/" is
mounted as read-only. Therefore, create mount point during init process before
"/" is remounted as read only. This is done in the same file as the definition
of coreu service.

Change-Id: If16431982a14663f8417673ee6744c9f8bd1c2d0
Signed-off-by: Yunwei Zhang <yunwei.zhang@intel.com>