[GPS] Re-factors Android properties setting - baytrail
BZ: 166914

By the past we needed to leverage on spid to set the gps properties value.
It was mainly to solve Clovertrail scaling.
As we do not meet anymore these constraints, we can now set these properties
from platform makefiles.
Objective is also to centralize GPS configuration at the same place.

Removing legacy artefacts as well.

Change-Id: If377843f1e20b3e1df78412c051dab1c68b60c2f
Signed-off-by: Nicolas SUET <nicolas.suet@intel.com>