[PORT FROM MAIN]Enable SurfaceFlinger to request a final CSC from the HWC in all cases.
BZ: 156459

For Google Miracast acceleration, The h264 encoder wants the source frames
to be delievered in NV12 format. This means that the HWC has to perform a
final CSC on any SF compositied frames as OGL cannot composite into NV12.

This compile time flag enables this behaviour in both the HWC and the
SurfaceFlinger. This patch is also needed to fix a CTS bug.

Change-Id: I5d58c4dec404db45b930bb1794d04106a4d6589d
Category: device enablement
Domain: Graphics-SF
Origin: internal
Upstream-Candidate: No
Signed-off-by: Gary Smith <gary.k.smith@intel.com>