iago: Fix the overriding of partition list
Product specific iago.ini has the partitions
overriding the base iago.ini file. Because of this, the
the plugins specific partition details are not reflected
in the final iago.ini file. This prevents creation of
plug-in partitions. For example: droidboot partition
which is included in engineering build is not visible
because of this overriding.

Change-Id: If91c9e6200ee91ebaf1349ac2904dd3a951ab0f9
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>