[AUDIO] Define minimum timeout for vibrator module - byt_t_ffrd8 -
BZ: 170100

Depending on the hardware type, vibrations under certain threshold cannot be
felt. This is mainly due to the vibrator slow response or the lag time.

To compensate this, the PFW now defines the minimum timeout allowed for a
short vibration. This parameter will be fetched by the Vibrator HAL and
forwarded to upper layers if necessary.

Change-Id: Iae21229f6c29dd89040693b32168453f90b78efc
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>