[AUDIO] Wrong order of mixer settings in PFW showing warnings in codec.
BZ: 168468

When capture or playback is getting muted, series of codec mixers get
updated. The order of mixer update is incorrect. Which sometimes lead to
warning on console. This will add 5 seconds latency to actually
turn off the codec.

In order to avoid that, last switch in the DAPM path should be turned
off first. This patch moves these switches to the Routing.Mute stage.

Change-Id: I8ee486c92c9b80f14e919f5b3d2b8cf57b202f59
Signed-off-by: lin <lin.a.zhuang@intel.com>