eMMC: Fix to mount partitions on eMMC device
This patch brings in the capability to boot on an
eMMC instead of a SATA disk. It should be noted that the image
can support either a SATA disk or an eMMC and not both at the
same time. With this patch, the device will boot with eMMC
only. The device cannot boot with SATA disk.

To boot with SATA disk again, revert this patch and create
a new image.

Change-Id: Ia9bb7f11ccb308b02a2fad20705881c812e412b2
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>
Signed-off-by: Mohanraj S <mohanrajx.sakkaraiappan@intel.com>