[PORT FROM R44B] slub: limit the max order to 2
BZ: 177364

As stated in document of kernel parameter, a high setting of
slub_max_order may cause OOMs due to memory fragmentation.

So in this patch, on CRv2 board with 1G memory, limit the
slub_max_order to 2 to reduce memory fragmentation.

Similar patch on previous LEX phone is as below:
https://android.intel.com/#/c/82643/

Change-Id: Ib5866b31323a9c17649dc1dae0aba6f5ec44cfe8
Signed-off-by: Li Fei <fei.li@intel.com>