Set property to disable the preventive graceful shutdown
BZ: 182823

Set a property to disable the preventive graceful shutdown in case
of long long press on power button.
That is to workaround a PMIC bug. If the power button press time
is shorter than forced shutdown time, but longer than graceful
shutdown, the device restarts, which is not the expected behavior.

Change-Id: I638efdd720c7df23fd1c609f1ce661303e6b7f78
Signed-off-by: Florent Auger <florent.auger@intel.com>