Remove redundant vibrator flags
BZ: 173061

The flag BOARD_USE_VIBRATOR is redundant with the product package
vibrator.$(TARGET_BOARD_PLATFORM) defined for each variant.

On the other hand, the PFW uses the property
Audiocomms.Vibrator.IsPresent to know if the variant has a vibrator.
This property is initialized to "true" by default.
Also, after removing the vibrator product package in a variant,
this property is no more used. So setting it to "false" is useless.

The goal of this patch is to remove:
 - the redundant flag BOARD_USE_VIBRATOR for all variants
 - the redundant/useless initialization of the property
   Audiocomms.Vibrator.IsPresent

Change-Id: Idbe100a57992e26523b832a937f4f9c738316aaf
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>