Remove DPTF startup.dv Configuration File from byt_m_crb Build
BMDA 481
android.security.cts.ListeningPortsTest
 - testNoRemotelyAccessibleListeningTcpPorts fail

The startup.dv file starts the DPTF web server that enables
OEM for DPTF tuning. However, this function would cause the
CTS to fail. Removing this file for official releases, and we will
give OEMs out-of-band instructions on how to enable the
DPTF web server.

Change-Id: I490dae4472e0ac07ed858285e697d94c5abde43d
Signed-off-by: Brian Bian <brian.bian@intel.com>
Reviewed-on: https://android.intel.com/178849
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Hiremath, Vijay P <vijay.p.hiremath@intel.com>
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>