DPTF:Fix device encrytion for the pid file creatd by dptf daemon in /data
Issue: BMDA-989
Tablet encrytion fails

Dptf daemon esif_ufd opens a PID lock file.
This is to ensure that multiple esif_ufd cannot run at the same time .
During encrytion when all the services are killed and /data 
unmounted vold was unable to unmount /data as the esif_ufd is still 
runiing and holding the lock file in /data/misc/dptf.

This patch changes the location of pid file from /data to /mnt

Change-Id: I705e132482b350d3501356d8dddd982f96cc6a53
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/181241
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>