baytrail: move the mPreviewFourcc from Common section to the Sensor section
BZ: 189953

Issue:
the preview format setting is set by platform right now,
it means that the preview format will be set to the same for front camera and back camera.
but the ISP output maybe different by sensors.

Solution:
we need to move the mPreviewFourcc from Common section to the Sensor section.

Change-Id: Id0bb7dba386eec55363598bcf9784c79e262a1b4
Signed-off-by: Yang Liang <liang.l.yang@intel.com>