fg_algo_ui_iface: Start fuel gauge algorithm in Charger OS
BZ: 201528

Start fuel gauge algorithm when Charger OS mode starts.
It is essential to start the FG algorithm in COS for
Baytrail CR V2.2 which uses Software Fuel Gauge and it
needs the user space algorithm to calculate the capacity,
charge_now, charge_full values and give it back to the
fule gauge driver.

Change-Id: I1413bbfba8f48744ea2ac526bac35981eac6be8b
Signed-off-by: raosrini <srinidhi.rao@intel.com>