ZhangIce: tuning accel wakeup threshold to 1/16g
borqsbtx: 61010
reviewers: xiaoxu.wang@borqs.com
rootcause: Accel will stop reporting data after a number of circles. The default
value of the circles is 120. Then, it goes into sleep and can be waken up by
motion detection. However, the default threshold of motion detection is too
small, and cause not easy to wake up.
This patch tunes the motion threshold.

Change-Id: I5be67978aee1d27fd92fdb3d7b67af83a5abb17c
Borig-Change-Id: I84f0548cdd6c8d38c1f56f31d2b410e15a644e99
Reviewed-on: http://192.168.1.79/gerrit/7180
Reviewed-by: b561 <b561@borqs.com>