byt_m_crb: liveimg: publish fastboot binary into the out/host folder
BMDA-1119

Build error is observed in the user build due to unavailability of
the fastboot binary in the out/host/linux-86/bin folder for the liveimg
build. This patch will copy the fastboot binary into the
out/host/linux-86/bin folder and resolve the user-build error.

This is an extension of the patch https://android.intel.com/190987/.

Change-Id: I4f37e556931d56e0b0f45d410d8e0486e332d55f
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>
Signed-off-by: Mohanraj S <mohanrajx.sakkaraiappan@intel.com>
Reviewed-on: https://android.intel.com/209190
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Krishnan, Ranganath <ranganath.krishnan@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>