byt_m_crb: Add ethernet settings
Issue: BMDA-1083

The original patch to enable internet over ethernet was 
https://android.intel.com/#/c/191472/.
That patch was reverted to resolve CTS failures as Wi-Fi and 
ethernet was not working together.
The revert patch was https://android.intel.com/#/c/201637/.
CTS issue is addressed in a different patch to make Wi-Fi and 
ethernet work together.

So we are re-submitting the patch to enable internet over ethernet.

Change-Id: I342d03edee355ebc9119a7d5a96781ce9baa315d
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/209583
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>