hwc: disabling panel fitting in HWC
borqsbtx: 61187
rootcause: Panel fitting is disabled when panel size is larger than 2k on BYT, but
currently there is no way to notice HWC. HWC team is working with Display team
for a solution on this.
A temporal solution could be disabling panel fitting in HWC if the panel
resolution is larger than 2k. I have confirmed with HWC team it should be low
risk to disable it.

Change-Id: I2157c3185680c1d9786d0f816e9aca58049e1558
Borig-Change-Id: I78eb5f3ffa47fbf0a5f3e2b1809e1e804e720e9c
Reviewed-on: http://192.168.1.79/gerrit/7590
Reviewed-by: b023 <b023@borqs.com>