[AUDIO] Apply modem "SPECIAL1" configuration for the PCM mode
BZ: 210183

Today, when probing the PCM signals during a BT CSV call,
the frame sync is 2 clock period wide
instead of 1 clock period wide.
This is causing some noise in BT headset.

As suggested by WPRD team, this patch will apply
the "SPECIAL1" configuration for the PCM mode.
For the I2S mode, we will continue to apply
the "NORMAL" configuration.

Change-Id: I5fbdbdc523e24f3db7a43815d2eb072f87b2b7dd
Signed-off-by: Neil Di Folco <neil.di.folco@intel.com>