[Audio] Remove BtEnable and ScoResourceRequested criteria
In iMin legacy, some bluetooth functionalities are missing. In particular,
there is no support for the criteria 'BtEnable' and 'ScoResourceRequested' at the route
manager level. The root cause is that the underlying functionalities are missing from
the audio libhardware (no parameter 'bluetooth_enabled' nor 'sco_resource_requested'.

This patch updates the routes.pfw file for baytrail projects (ffrd8, crv2): now the rules
allowing BtComm and BtCSV to be applicable in capture and playback do not take into account
the criteria BtEnable or ScoResourceRequested.

Change-Id: Ide32a38e6cb72f450d2d5fe47a9963c40afcff97
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1326
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: https://android.intel.com:443/234118