Disable MDS and VPP temporarily
BZ: 220484

Currently, MDS starting code isn't ported to framework,
it means MDS isn't started, and will cause caller fails to
check MDS service, and cause some video playback errors.
So, we disable it temporarily,
and will enable it after MDS framework hack code is approved.

Disalbe VPP on Baytrail as it is not enabled on Imin-Legacy.

Change-Id: I515fb55266578fa756f9f04474cbe7bba2e8a1ef
Signed-off-by: Yuanjun Huang <yuanjun.huang@intel.com>
Reviewed-on: https://android.intel.com:443/236206