device_baytrail: Add new tags to camera_profiles.xml
Now that HAL block of the CPF is being removed,
this patch adds tags to camera_profiles.xml that
used to be in HAL block.

byt_t_ffrd8 values taken from 00imx134.hal and 01ov2722.hal
byt_t_crv2 values taken from 00ov5693.hal and 01ov9724.hal

This patch changes also the device.mk files to match
the new file name (.aiqb instead of .cpf).

It was agreed with the CTS team that the anzhen related files
are not updated.

Change-Id: I650b27d8248b46f9ece3b05b96675be9fa0aba84
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3812
Signed-off-by: Kaisa Taipale <kaisa.taipale@intel.com>
Reviewed-on: https://android.intel.com:443/285269