Add config to turn off global scaling
The property is to turn off global scaling from HW compositor. With global
scaling on, if a display is 1920x1080 or 1920x1200 and video is 720p, HWC will
use display panel fitter to scale up.  But, currently, there is an issue upscale
to 1920x1080 monitor.

Since there isn't a big performance gain with global scaling turn on, will turn
it off now until HWC has the issue resolve.

Sighting: 4994810 and 4994741

Change-Id: I6ccf6aec35e0c5cda68b135016eecbf3df46290d
Signed-off-by: Chon Ming Lee <chon.ming.lee@intel.com>
Reviewed-on: https://android.intel.com/289004
Reviewed-by: Bodke, Kishore K <kishore.k.bodke@intel.com>
Tested-by: Bodke, Kishore K <kishore.k.bodke@intel.com>