[AUDIO] Remove id from all route subsystem component
The route manager has been changed to avoid the plugin giving an index
when adding components as port and routes. The route manager will
assign an index when creating the associated criteria.
The route plugin has been changed accordingly, so it is not required
anymore to add the id as a mapping information for these components.

This patch removes the id mapping field to all route subsystem
component (Route / Port / Port group).

Change-Id: I5ed60bf31d53bb5e6089c38d4fa8dcf4abf9d6b4
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6999
Reviewed-on: https://android.intel.com:443/297803