[AUDIO] Declare CSV route as full duplex route
The CSV route is declared as playback only.
With the refactor of the route plugin structure, the route mgr
is no more permissive to error declaration. As a result, the rules
on CSV route criterion in capture is failing.

This patch declares the CSV route as a full duplex as it shoudl have
been.

Change-Id: I07dbaa93148a13cad929aa7d63f1a93f6a6f665e
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6999
Reviewed-on: https://android.intel.com:443/301472