baytrail: MMGR - class modification
MMGR service used to be configure as a core service in order to
start the modem at the earliest. This is not required anymore as
the modem will be started only when a client acquires the
resource.

Plus, MMGR uses persistent properties to detect a modification of
its configuration. Those properties are not accessible while the
data partition is encrypted. That is why, MMGR needs to be
restarted, like the RIL, once the partition is decrypted.

Configuring MMGR as a main service instead of a core one is
more relevant for all those reasons.

Change-Id: I60aab9f08e75b32d6bd3591b7962924049784344
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10633
Signed-off-by: Cesar DE OLIVEIRA <cesar.de.oliveira@intel.com>
Reviewed-on: https://android.intel.com:443/306299