byt_m_crb: audio: remove id from route subsystem component
The route manager has been changed to avoid the plugin giving an index
when adding components as port and routes. The route manager will
assign an index when creating the associated criteria.
The route plugin has been changed accordingly, so it is not required
anymore to add the id as a mapping information for these components.

This patch removes the id mapping field to all route subsystem
component (Route / Port / Port group).

Change-Id: Idf89bc9cff947920575e15e2d05d3637b84d8914
Signed-off-by: Ranganath Krishnan <ranganath.krishnan@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6999
Reviewed-on: https://android.intel.com:443/297803
Reviewed-on: https://android.intel.com/308022
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>
Tested-by: Kandru, Suneel <suneel.kandru@intel.com>
Reviewed-by: Kankroliwala, Huzefa NomanX <huzefa.nomanx.kankroliwala@intel.com>